Better error messages for bad setter naems
The error message for invalid properties includes the property name. If the error was an invalid property name on a method setter, the resulting error would indicate the problem was with property named 'null', which wasn't particularly helpful, although it was somewhat more precise. Rather than changing the way the message is written, this change simply adopts the whole method name as the property name if the property name is invalid.