junit.framework -> org.junit
Replaced all of the uses or junit.framework.* with org.junit.*; this is long overdue. In
the process, I also replaced deprecated assertEquals(double,double) with
assertThat( double, is( double ) ) because these aren't calculated doubles,
and I don't wany any epsilon of difference, or to have to @SurpressWarnings
the deprecation.

No associated github issue.