Cleaning Up Left-Over 'Origin' References
When I first broke source properties into a class, I called that class 'Origin',
then ended up renaming it to 'SourceProperty' before I checked it in. Some of
the variables and methods that were related to 'origins' did not get renamed
to 'source properties' along the way, so this is cleaning up some remaining
variables and methods as I was making some other changes.

Related to #14.