Tests for Single-Character Properties
Added two tests for the ReflectionSourceProperty to ensure that it works with single-character
fields and methods, since these were an edge case that I'd written down during the implementation
as a potential concern.

They worked, so this is additional test code with no changes to the reflection source property.

Related to #14.